requirements-railway.txt
# Optimized requirements for Railway (CPU-only, no GPU)
python-telegram-bot>=22.5
sahi>=0.11.18
opencv-python-headless>=4.8.0
ultralytics>=8.0.0
# CPU-only PyTorch (much smaller and faster to install)
--index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
python-dotenv>=1.0.0




